+++ changelog.md
@@ -1,12 +1,18 @@
 # changeglob
 
+## 12
+
+- Remove the unsafe `--shell` option. The `--shell` option is now
+  ONLY supported on known shells where the behavior can be
+  implemented safely.
+
 ## 11.1
 
 [GHSA-5j98-mcp5-4vw2](https://github.com/isaacs/node-glob/security/advisories/GHSA-5j98-mcp5-4vw2)
 
 - Add the `--shell` option for the command line, with a warning
   that this is unsafe. (It will be removed in v12.)
-- Add the `--cmd-arg`/`-g` as a way to *safely* add positional
+- Add the `--cmd-arg`/`-g` as a way to _safely_ add positional
   arguments to the command provided to the CLI tool.
 - Detect commands with space or quote characters on known shells,
   and pass positional arguments to them safely, avoiding